libsandbox: let real funcs handle non-existent paths
If no explicit deny exists on a path, and some leading components of the
path do not actually exist, pass the path on to the actual function when
we know said functions will fail.  For example, many tests operate on some
kind of path where the leading components do not exist.  Such as the m4
test that tries to do fopen("/none/such", "w") and the /none/ dir does not
actually exist.  Since we know /none/ does not exist, let through funcs
that will fail in this case.

URL: http://bugs.gentoo.org/244396
Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Reported-by: Ryan Hill <dirtyepic@gentoo.org>