libsandbox: handle ENAMETOOLONG properly with getcwd()
The core libsandbox code was checking for ENAMETOOLONG when working with
getcwd() and doing the right thing, but it never passed this error back up
to the caller.  This would make some tests (like the getcwd tests in many
gnulib/autotool based packages like coreutils) to keep on trying since it
did not detect the expected failure conditions.  So now when we hit the
ENAMETOOLONG error, we pass this specific error back up to the caller.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>