libsandbox: save/restore errno for hardened
Since hardened systems may return -1 in the case of an error, but this
isn't a "real" error, we need to save/restore errno.  Otherwise we corrupt
the errno value of the caller.  This is a regression on hardened systems
due to the recent commit bab59e2c which optimized symbol loading a bit.

URL: http://bugs.gentoo.org/260765
Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Reported-by: Magnus Granberg <zorry@ume.nu>