tests: convert `touch` to `open-0`
Convert the `touch` test to use our `open-0` util so we get fine grained
control over the test and so we don't have to worry about localization
issues or about system-specific messages.

This also extends the test framework to start parsing symbolic open()
flags as well as errno values (and add support for checking them).

Signed-off-by: Mike Frysinger <vapier@gentoo.org>