libsandbox: handle long paths with *at funcs
If a path is too long, then attempting to read the fd link in /proc/*/fd/#
will fail with ENAMETOOLONG.  In that case, we should restore errno and
let the normal function do its business.  Chances are we wouldn't have
gotten this far in the first place if there was an actual deny in place.
This fixes some of the testcases in coreutils that create really long
paths and then test things.

URL: http://bugs.gentoo.org/259876
Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Reported-by: Ryan Hill <dirtyepic@gentoo.org>