libsandbox: handle unreadable dirs better as non-root
If attempts to access the specified path fails, this is most likely due to
funky perms being setup.  Since this is really only an issue for non-root
users when testing stuff (like the coreutils tests), we can let the real
func do it's work when running as non-root.  If the path contains links to
some place we would normally deny, then attempt to dereference what little
we can.  If all else fails, the standard unix permissions will prevent
damage from being done even if we are unable to log it.

URL: http://bugs.gentoo.org/259876
Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Reported-by: Ryan Hill <dirtyepic@gentoo.org>