libsandbox: don't filter signals the traced pid receives
If we receive a notice that the child got a signal we don't care about,
make sure we tell it to continue on with the signal info so we don't go
filtering all signals the child may receive.  Otherwise we break test code
like that in glibc which exercises the ability of a child to catch and
process signals properly.

URL: http://bugs.gentoo.org/265072
Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Reported-by: Nick Fortino <nfortino@gmail.com>