libsandbox: don't swallow SIGCHLD notifications
When tracing static processes, the original implementation included code
that would always swallow SIGCHLD.  Much has changed since then, and it
doesn't seem to be needed anymore, and it is certainly breaking a few
packages.  So drop it, add some tests, and if it causes a regression in
the future, we can look at it then (with an actual test case).

URL: http://bugs.gentoo.org/289963
Reported-by: Joeri Capens <joeri@capens.net>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>