libsandbox: avoid passing same buffer to erealpath
The erealpath function modifies the storage buffer given to it in place
and can misbehave if both the source and destination buffers point to the
same storage in memory.  So fix the one case where we were doing this in
the canonicalize() function and add some run time checks to make sure this
doesn't crop up again.

URL: http://bugs.gentoo.org/292050
Reported-by: Hongjiu Zhang <voidprayer@gmail.com>
Reported-by: Fredric Johansson <johansson_fredric@hotmail.com>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>