tests: force a saner default env
Rather than having tests rely on implicit behavior (current dir is not
listed anywhere thus it is denied), force tests to declare explicit
deny paths when desired.  This way `make check` works even when it is
run under a path that is granted write access by default (such as the
tempdirs).

Signed-off-by: Mike Frysinger <vapier@gentoo.org>