tests: addpredict / by default to match portage
Portage by default will do `addpredict /` which can causes tests that
expect certain behavior to fail, but only when testing under portage.
So tweak the default environment to include this so that we are forced
to tweak the tests so that they pass in and out of the portage env.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>