libsandbox: mark rmdir as working on symlinks
Doing rmdir(symlink) does not remove the dir the symlink points to, but
will operate on the symlink itself.  While it will always fail (since it
is a link and not a dir), that isn't something we need worry about.  Just
need to avoid doing permission checking on the target of the symlink.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>