libsandbox: fix bug in previous dirfd unification
The previous commit (libsandbox: handle dirfd in mkdir/open/unlink *at
prechecks) left a sizeof() in place but unfortunately no longer held the
same meaning.  In previous code, the function had access to the buffer
decl and so could get the byte count.  In the new code, the function has
access to the pointer only.  So sizeof() now wrongly returns the size of
pointers rather than the length of the buffer.

Extend the new helper function to take the length of the buffer it is
given to fix this issue.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>