tests: always bubble up exit code regardless of log existence
To simplify testing, do not mung exit(0) to exit(1) just because the log
file exists.  In many of our tests, we will be doing things to generate
a log file, but we explicitly test for exit values ourselves.

This is also needed to make log file handling more resilient where we
get the name at startup, but don't allow live env changes after that.
The changing of the log name to sb.log on the fly no longer works.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>