sb_printf: expand feature set slightly and fix testing
This adds support for signed ll, unsigned z, l, and ll, hex l, and ll,
ignores the # for hex output since this is what we do implicitly already.

As for testing, looks like during the autogeneration of testsuite.list.at,
the sb_printf test was lost.  Restore it so it gets run again.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>