sb_gdb: improve gdb integration
Add a dedicated entry point for connecting gdb to make it easy to connect
gdb at arbitrary points (ala printf style debugging).

This also smooths a lot of the common steps when automatically launching
gdb such as making sure the process is closer to the crash point when the
user takes over control of gdb.

Finally, switch to using clone rather than fork since the latter relies
on the C lib's fork which implicitly can grab locks.  If we're crashing
in the middle of a func that already holds those locks, the fork call
will hang indefinitely on us.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>