libsandbox: fall back to tracing set*id programs
If we are non-root and run a set*id program, the ldso will ignore our
LD_PRELOAD (rightly so).  Unfortunately, this opens up the ability to
run set*id apps that modify things and sandbox cannot catch it.

Instead, force ptracing of these ELFs.  While the kernel will disallow
the set*id aspect when running, for the most part, that shouldn't be a
problem if it was already safe.

URL: http://bugs.gentoo.org/442172
Reported-by: Nikoli <nikoli@lavabit.com>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>