sandbox: use a non-shell var for the message path
By allowing the SANDBOX_MESSAGE_PATH var to be stored in the shell
environment and then modified on the fly, we run into a fun edge
case with the PM.  When a phase has finished running, it saves the
current environment.  When the next phase runs, it loads the env
from the previous run.  Since the message path var can contain a
pid, the previous run will no longer be valid.

Since we want this to simply be a way for the active sandbox to
pass information to the active libsandbox.so's, there's no need
to use an env var that the shell can save/reload.  As such, use a
variable name that the shell will skip.  Non-shell programs have
no problem with this.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>