libsbutil: undef memory redirect calls
Sometimes the C library will redirect a call to strdup to __strdup which
breaks when we're using the libsandbox memory allocator.  This was fixed
in libsandbox in commit d7801453aced46a6f31d8455877edeb31a5211cc, but we
didn't notice in libsbutil as no calls to strdup happened to come up.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>