libsandbox: work around process_vm_readv EFAULTs
Some people are seeing this call fail, but it's not clear why.  Include
more debugging output so as to improve the reports, and let the code fall
back to the existing ptrace logic since that seems to work.  This will at
least unblock people's builds.

URL: https://bugs.gentoo.org/560396
Reported-by: Jeroen Roovers <jer@gentoo.org>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>