libsandbox: do not abort when the target uses bad pointers
If the target passes a bad pointer to the kernel, then trying to extract
the data via ptrace will also throw an error.  The tracing code should not
abort though as there's no valid address to check, and kernel itself will
return an error for us.  Simply return and move on.

URL: https://bugs.gentoo.org/560396
Reported-by: Jeroen Roovers <jer@gentoo.org>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>