libsandbox: tweak edge cases of realloc a bit
We need to return NULL when passed a size of 0 as the API requires the
return value be usable w/free, but we just freed the pointer so the ret
will cause memory corruption later on.

When we go to preserve the old content, we don't need the MIN check as
we already verified that a few lines up.  But leave it for defensive
purposes as gcc already optimizes it out for us.  Just comment things.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>