build: fix SB_SCHIZO automake conditional
The rework in commit 46fe624223cfe62fb6c2fbb609be42f2f1d1734b broke the
set up of the SB_SCHIZO automake conditional for non-schizo builds as it
was not updated to the new variable.  This would cause the syscall table
to always be empty and thus the ptrace code would never match.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>