libsandbox: Remove meaningless/broken -nodefaultlibs
Remove '-nodefaultlibs' from linking flags for libsandbox as it is
apparently meaningless and broken at the same time.  When regular
libtool is used, it silently strips the option, making it meaningless.
When slibtool is used instead, it passes the option which causes linking
to fail due to undefined symbols.

Thanks to the bug reporter and slibtool devs from researching
the problem in detail.

Bug: https://bugs.gentoo.org/657184