tests: disable utimensat-3 on *-linux-musl
x86_64-gentoo-linux-musl fails a single test:
    83: utimensat/3     FAILED (utimensat.at:3)

The test checks if sandbox does not crash when
    utimensat(<filefd>, NULL, NULL, 0)
is called. The behaviour is not specified by POSIX
but glibc returns EINVAL for such a case. Thus the
test behaves differently on varius libs.

https://www.openwall.com/lists/musl/2019/06/25/1 has
a conversation with musl upstream.

The change restricts test down to glibc targets.

Bug: https://bugs.gentoo.org/549108
Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>