libsandbox: implement vfork() via fork()
sandbox turns
    vfork()/exec("/sbin/ldconfig")
into
    vfork()/ptrace()+fork()/exec("/sbin/ldconfig").

It happens because "/sbin/ldconfig" is a static binary and can't be
inspected via LD_PRELOAD and sandbox falls back to fork()+ptrace()

vfork() imposes very strong requirements on what could happen between
vfork() and exec(). Above sandbox behaviour violates it.

vfork() is specified in a way that it can always can be substituted
for fork(). This change does exactly that.

Reported-by: Michał Górny
Bug: https://bugs.gentoo.org/774054
Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>