libsandbox: enable exitkill containment
If the ptrace code crashes/aborts for any reason, don't let the tracee
leak out.  No one will be expecting this, and it could lead to zombie
processes hanging around.

This option is new to Linux 3.8.  We could code this so that it would
fallback automatically to older versions of Linux, but with 3.8 being
released in 2013, doesn't seem like we need to support anyone that old.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>