libsandbox: use wide readelf output
Newer versions of binutils will truncate symbol output weirdly unless
the --wide option is used.  This manifests itself as libsandbox not
including symbols when their name and symbol version is too long.
The new removexattr symbol tests were failing because of this, but
it seems the others were either not too long, or we didn't have any
test coverage for them (oops).

Signed-off-by: Mike Frysinger <vapier@gentoo.org>