libsandbox: fix flags extraction for a few syscalls
While many syscalls follow similar patterns for dirfd & path handling,
the flags argument is less consistent -- it tends to be last with all
other arguments in between.  As a result, a few syscalls were pulling
the wrong argument for the flags settings:
* fchmodat: the syscall interface has no flags at all
* fchownat: the flags come after uid & gid
* utimensat: the flags come after the timespec

These syscalls haven't been a problem in practice because no one ever
tries to chmod/chown/utimes on symlinks themselves.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>