scripts: rewrite main processing loops for significant speedup
The awk scripts iterate over all the possible symbols for each line of
readelf output.  As we add more symbols, and as the C library grows,
the number of iterations explodes.

We iterate over the list of possible symbols by creating a regex to
match against the readelf output.  We could create a large regex at
the start of the script to match all possible symbols, and then run
that against the readelf lines.  This avoids the nested loop logic,
and speeds up the scripts significantly: from ~1.5sec to ~0.05sec.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>