libsandbox: move symbols.h.in parsing to scripts
In preparation for extending the symbol format, move parsing out of
the makefile (which is a basic sed expression) to the awk scripts.
This also has a nice side benefit of removing one automake warning.

It is slightly more code, but the scripts will be diverging shortly,
so it's unavoidable.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>