libsandbox: tweak how undefined symbols are declared
Rather than always set undefined symbols to the same constant, expand
it to a range of constants, and give every symbol a unique value.  For
dynamic symbol processing, this isn't a big deal as such symbols will
never show up, but when handling syscalls that don't have a matching C
library symbol, we need to make sure that we have unique entries.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>