sandbox: change interface to make it easier to pass thru
The sandbox command line is passed to a shell for execution.  This can
be a bit awkward to quote right if you weren't expecting it, and even
if you were.  Change the default behavior to be more like `env` where
the arguments, as they are, get passed through and run.  If people want
the old shell behavior, they can use the -c option akin to `bash -c`.

Bug: https://bugs.gentoo.org/265907
Signed-off-by: Mike Frysinger <vapier@gentoo.org>