sandbox: add backwards compat interface hack
Portage runs commands through sandbox like:
  $ sandbox "/usr/lib/portage/python3.9/ebuild.sh unpack"

That means we can't break the CLI without breaking portage and forcing
everyone to upgrade together.  That'll be pretty disruptive for people,
so add a hack to detect this situation: if a single argument is passed
on the CLI, and it doesn't appear to be a file, then fallback to running
it through the shell.  This keeps portage working while allowing the new
interface style to launch.  If/when we can update portage to always use
the -c option, maybe we can drop this in the future.  Or not ... it's
not exactly the worst hack for users.

Bug: https://bugs.gentoo.org/265907
Signed-off-by: Mike Frysinger <vapier@gentoo.org>