libsandbox: merge sandbox settings from tracee end when execing
This allows traced children to change their sandox settings on the fly
and the out-of-process tracer will react accordingly.  We don't try to
read the environ all the time as it's kind of impossible to know where
the tracee is storing it (since it can point |environ| anywhere).

This means turning the sandbox on/off won't work in the current process,
only in forked children.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>