libsandbox: add YAMA checks and skip ptrace when active
The YAMA ptrace_scope knob restricts access to different ptrace calls
depending on the capabilities the current process holds.  For now, do
not try to ptrace processes when the YAMA level is incompatible with
the capabilities that we have.

This means we basically cannot protect against processes when they
get into this state, so for now, we release them rather than abort.

Bug: https://bugs.gentoo.org/771360
Bug: https://bugs.gentoo.org/821403
Signed-off-by: Mike Frysinger <vapier@gentoo.org>