libsandbox: reduce & inline the __64_{pre,post}.h headers
Now that we use 64-bit stat & lstat explicitly everywhere, we don't
need these dynamic redirects for 64-bit wrappers.  The off_t define
is only used by one file anymore too, but we can inline that.

That leaves the SB64 define which we use inconsistently in places.
In some 64-bit modules that include the 32-bit, we use SB64 to switch
between the 64-bit & 32-bit APIs.  In other places, the 64-bit file
is responsible for redefining the few relevant APIs.  Let's switch
all the files away from SB64 and to defining the single thing that
the 64-bit module needs directly.  It's either the same or fewer LOC
this way, and doesn't seem any more or less difficult to maintain.
The __64_{pre,post}.h & SB64 define weren't easily discoverable.

Bug: https://bugs.gentoo.org/583282
Signed-off-by: Mike Frysinger <vapier@gentoo.org>